requirements.txt
# NEPSE AutoScan — Python dependencies
# Install: pip install -r requirements.txt
#
# GPU packages (install separately):
#   pip install torch --index-url https://download.pytorch.org/whl/cu124
#   pip install xgboost lightgbm  (CUDA support built-in)

# ── Core ──────────────────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0              # GPU: device='cuda', tree_method='hist'
lightgbm>=4.0.0             # GPU: device='gpu'
# torch>=2.0.0              # Install separately with CUDA index (see above)
# hmmlearn>=0.3.0           # Optional: HMM regime detection

# ── Data & Web ────────────────────────────────────────────────────────────────
requests>=2.28.0
beautifulsoup4>=4.9.0       # HTML parsing for Sharesansar scraping
python-dotenv>=0.20.0       # .env credential management
tabulate>=0.9.0             # Report table formatting

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.0.0

# ── LLM (local, no pip install needed) ────────────────────────────────────────
# Ollama with Qwen 2.5 14B: https://ollama.ai/
#   curl -fsSL https://ollama.ai/install.sh | sh
#   ollama pull qwen2.5:14b
